Fix #3 by printing to tty and not stdout
When you do something like

>  confirm-before echo "fooo" | grep foo

Nothing was happening (apparently).

Under the hood the grep commands consumes the stdout (where we print the
"Are you sure?(y/n)"), then it keeps waiting for the stdin flush.

Therefore, we need to write to tty instead of writing to stdout, otherwise
grep suppress the output... it waits silently for the stdin input.

We use termion to do that, simple and works fine.

Unfortunately, I couldn't find an easy way to test for that... I accept
suggestions.